all real variables are now real(8), of double precision

+++ opo/tophat/defect/RK_OPO_rk.f90
@@ -4,8 +4,8 @@
         USE rk_adaptive
         IMPLICIT NONE
 
-        REAL :: x1_r,x2_r,h1_r,hmin_r
-        REAL, EXTERNAL ::  findfermpart,findcoopernum
+        real(8) :: x1_r,x2_r,h1_r,hmin_r
+        real(8), EXTERNAL ::  findfermpart,findcoopernum
         COMPLEX, EXTERNAL :: findfermcond1, findfermcond2
 
         call read
