+++ opo/tophat/defect/RK_OPO_sp.f90
@@ -5,8 +5,8 @@
         USE FFTW3
         IMPLICIT NONE
 
-        REAL :: x1_r,x2_r,h1_r,hmin_r
-        REAL, EXTERNAL ::  findfermpart,findcoopernum
+        real(8) :: x1_r,x2_r,h1_r,hmin_r
+        real(8), EXTERNAL ::  findfermpart,findcoopernum
         COMPLEX, EXTERNAL :: findfermcond1, findfermcond2
 
         ! fft stuff
