+++ opo/tophat/defect/ode_path.f90
@@ -2,6 +2,6 @@ Module ode_path
   USE nrtype
   INTEGER(I4B) :: nok,nbad,kount
   LOGICAL(LGT), SAVE :: save_steps=.true.
-  REAL :: dxsav_rk, dxsav_sp
+  real(8) :: dxsav_rk, dxsav_sp
 end Module ode_path
